% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scores.R
\name{calc_scores}
\alias{calc_scores}
\title{Calculate various scoring rules for a Time Series Dataset}
\usage{
calc_scores(dataset, parallel = FALSE, num.cores = 2)
}
\arguments{
\item{dataset}{the list containing the series. See details for the required format.}

\item{parallel}{logical. If \code{TRUE} then the calculations are conducted in parallel.}

\item{num.cores}{the specified amount of parallel processes to be used if parallel = TRUE.}
}
\value{
A list with the elements having the following structure
\describe{
  \item{MASE}{a matrix with \code{F} rows and \code{h} columns. Each row contains 
  the MASE scores of each method in \code{methods}.
  Each column represents the MASE for the j-step point forecasts, where j=1,...,h. }
  \item{sMAPE}{a matrix with \code{F} rows and \code{h} columns. Each element 
  in the matrix denotes the sMAPE values.}
  \item{MSIS}{a list with each element being a matrix for certain confidence level in 
  \code{level}. The matrix contains the MSIS values in corresponding forecasting 
  horizon for each method.}
  \item{Spread}{a list with each element being a matrix for certain confidence level in 
  \code{level}. The matrix contains the Spread values in corresponding forecasting 
  horizon for each method.}
  \item{IfInn}{a list with each element being a matrix for certain confidence level in 
  \code{level}. Each element in the matrix measures if the true value lies inside 
  the prediction interval.}
  \item{IfInu}{a list with each element being a matrix for certain confidence level in 
  \code{level}. Each element in the matrix measures if the true value is not larger than 
  the upper bound of the prediction interval.}
}
}
\description{
For each series in \code{dataset}, point and interval forecasting performance 
are evaluated for all methods in \code{methods} of \code{\link{ts_forec}} 
with regard to MASE, sMAPE, MSIS, Spread, Coverage and Upper coverage.
}
\details{
\code{dataset} must be a list with each element having the following format:
\describe{
  \item{x}{a time series object \code{ts} with the historical data.}
  \item{h}{the amount of future time steps to forecast.}
  \item{f}{a matrix with \code{F} rows and \code{n} columns. Each row contains
  the fitted values of each method in \code{methods}.}
  \item{ff}{a matrix with \code{F} rows and \code{h} columns. Each row contains
  the forecasts of each method in \code{methods}.}
  \item{lower}{a list with each element being the matrix of lower bounds 
  for certain confidence level.}
  \item{upper}{a list with each element being the matrix of upper bounds 
  for certain confidence level.}
}
}
