% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.threshold.R
\name{choose.threshold}
\alias{choose.threshold}
\title{Search the optimal threshold ratio used for model averaging}
\usage{
choose.threshold(
  dataset,
  fitProb,
  ratio,
  combine = "mean",
  level,
  parallel = FALSE,
  num.cores = 2
)
}
\arguments{
\item{dataset}{a list of time series. See details for the required format.}

\item{fitProb}{a matrix with \code{n} (number of time series) rows and 
\code{F} (number of benchmark methods) matrix. The matrix is applied to assign weights 
to each benchmark method for each time series.}

\item{ratio}{a vector with pre-set threshold ratios. The threshold ratio 
quantifies as a number between 0 and 1, where 0 indicates all the benchmark methods 
are selected and 1 indicates only the benchmark method with the minimal value of 
scores is selected.}

\item{combine}{string representing the interval combination methods. 
"mean", "weightedmean" or "median".}

\item{level}{the confidence level of prediction intervals that are used for combination.}

\item{parallel}{logical. If \code{TRUE} then the calculations are conducted in parallel.}

\item{num.cores}{the specified amount of parallel processes to be used if parallel = TRUE.}
}
\value{
A list with the following elements
\describe{
  \item{threshold_y}{numeric, the optimal threshold ratio searched for yearly data.}
  \item{threshold_q}{numeric, the optimal threshold ratio searched for quarterly data.}
  \item{threshold_m}{numeric, the optimal threshold ratio searched for monthly data.}
  \item{msis_y}{a matrix of MSIS values for each pre-set threshold ratio to yearly data.}
  \item{msis_q}{a matrix of MSIS values for each pre-set threshold ratio to quarterly data.}
  \item{msis_m}{a matrix of MSIS values for each pre-set threshold ratio to monthly data.}
}
}
\description{
According to the fitted values of the pre-trained algorithm, 
the optimal threshold ratio used for model averaging is searched 
from a set of pre-set threshold ratios.
}
\details{
\code{dataset} must be a list with each element having the following format:
\describe{
  \item{n}{the number of observations in the time series.}
  \item{h}{the number of required forecasts.}
  \item{period}{interval of the time series. Possible values are "YEARLY", "QUARTERLY", "MONTHLY" & "OTHER".}
  \item{x}{a time series of length \code{n} (the historical data).}
  \item{xx}{a time series of length \code{h} (the future data).}
  \item{ff}{a matrix with \code{F} rows and \code{h} columns. Each row contains
  the forecasts of each method in \code{methods}.}
  \item{lower}{a list with each element being the matrix of lower bounds 
  for certain confidence level.}
  \item{upper}{a list with each element being the matrix of upper bounds 
  for certain confidence level.}
}
}
