% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_msis.R
\name{extract_msis}
\alias{extract_msis}
\title{Extract MSIS values for certain confidence level from a list}
\usage{
extract_msis(dataset, level)
}
\arguments{
\item{dataset}{A list with each element having a \code{MSIS} fields.}

\item{level}{the confidence level for which MSIS values are extracted.}
}
\value{
A matrix with \code{N} rows and \code{F} columns. Each row contains
  the MSIS values of each method in \code{methods}.
}
\description{
Extract MSIS values for certain confidence level from a list 
that each element contains a list named as \code{MSIS}.
}
