% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_model.R
\name{gam.fun}
\alias{gam.fun}
\title{Train generalized additive models}
\usage{
gam.fun(X, Y, LogY = TRUE, k = 10, parallel = FALSE, num.cores = 2)
}
\arguments{
\item{X}{a dataframe with \code{n} rows and \code{c} columns, 
where \code{n} is the number of observations and 
\code{c} is the number of explanatory variables.}

\item{Y}{a dataframe with \code{n} rows and \code{F} columns. 
Each column is considered as a response variable to fit the generalized additive model.}

\item{LogY}{logical. If \code{TRUE} then the logarithmic form of \code{Y} 
becomes the response variable.}

\item{k}{the dimension of the basis used to represent the smooth term 
when \code{mgcv} package is used to fit GAM. 
If \code{k} is not specified then basis specific defaults are used.}

\item{parallel}{logical. If \code{TRUE} then the model training is conducted in parallel.}

\item{num.cores}{the specified amount of parallel processes to be used if parallel = TRUE.}
}
\value{
\code{gam.fun} returns a list with \code{F} generalized additive models 
for each response variable.
}
\description{
Fit a generalized additive model(GAM) for data to capture 
the relationship between the relationship between the X and each column of Y.
}
\examples{
X <- as.data.frame(state.x77[, "Murder"])
Y <- as.data.frame(state.x77[, c("Population", "Illiteracy", "Income", "Frost")])
gam.fit <- gam.fun(X, Y, LogY = FALSE, k = 10)
}
