% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forec_methods.R
\name{methods_list}
\alias{methods_list}
\alias{snaive_fun}
\alias{naive_fun}
\alias{auto_arima_fun}
\alias{ets_fun}
\alias{tbats_fun}
\alias{stlm_ar_fun}
\alias{rw_drift_fun}
\alias{thetaf_fun}
\title{Benchmark methods list}
\usage{
methods_list()

snaive_fun(x, h, level)

naive_fun(x, h, level)

auto_arima_fun(x, h, level)

ets_fun(x, h, level)

tbats_fun(x, h, level)

stlm_ar_fun(x, h, level)

rw_drift_fun(x, h, level)

thetaf_fun(x, h, level)
}
\arguments{
\item{x}{a \code{ts} object with the input time series}

\item{h}{the amount of future time steps to forecast}

\item{level}{the confidence levels for prediction intervals.}
}
\description{
A list of the forecasting methods used for calculating 
the point forecasts and prediction intervals with different confidence levels.
}
\section{Functions}{
\itemize{
\item \code{snaive_fun}: forecast::snaive

\item \code{naive_fun}: forecast::naive

\item \code{auto_arima_fun}: forecast::auto.arima

\item \code{ets_fun}: forecast::ets

\item \code{tbats_fun}: forecast::tbats

\item \code{stlm_ar_fun}: forecast::stlm with ar modelfunction

\item \code{rw_drift_fun}: forecast::rwf

\item \code{thetaf_fun}: forecast::thetaf
}}

