% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simp_forec.R
\name{simp_forec}
\alias{simp_forec}
\title{Combinate the forecasts and prediction intervals by simple average method}
\usage{
simp_forec(
  dataset,
  level,
  method.num,
  methodname = "comb",
  parallel = FALSE,
  num.cores = 2
)
}
\arguments{
\item{dataset}{a list of time series. See details for the required format.}

\item{level}{the confidence level of prediction intervals that are used for combination.}

\item{method.num}{the number of methods used for averaging. Methods in the first 
\code{method.num} rows in \code{ff} will be selected for simple averaging.}

\item{methodname}{string. the name used to name the combined forecasting results.}

\item{parallel}{logical. If \code{TRUE} then the calculations are conducted in parallel.}

\item{num.cores}{the specified amount of parallel processes to be used if parallel = TRUE.}
}
\value{
A list with the elements having the following structure
\describe{
  \item{ff}{a matrix with \code{F+1} rows and \code{h} columns where 
  the first \code{F} rows represent the point forecasts of benchmark methods 
  and the last row represents the point forecasts by model averaging.}
  \item{lower}{a list with one element that contains a matrix. The matrix with 
  \code{F+1} rows and \code{h} columns where the first \code{F} rows represent 
  the lower bounds of benchmark methods and the last row represents the 
  lower bounds by model averaging.}
  \item{upper}{a list with one element that contains a matrix. The matrix with 
  \code{F+1} rows and \code{h} columns where the first \code{F} rows represent 
  the upper bounds of benchmark methods and the last row represents the 
  upper bounds by model averaging.}
}
}
\description{
The combinated forecasts and prediction intervals are calculated 
by simple average method.
}
\details{
\code{dataset} must be a list with each element having the following format:
\describe{
  \item{n}{the number of observations in the time series.}
  \item{h}{the number of required forecasts.}
  \item{period}{interval of the time series. Possible values are 
  "YEARLY", "QUARTERLY", "MONTHLY" & "OTHER".}
  \item{x}{a time series of length \code{n} (the historical data).}
  \item{xx}{a time series of length \code{h} (the future data).}
  \item{ff}{a matrix with \code{F} rows and \code{h} columns. Each row contains
  the forecasts of each method in \code{methods}.}
  \item{lower}{a list with each element being the matrix of lower bounds 
  for certain confidence level.}
  \item{upper}{a list with each element being the matrix of upper bounds 
  for certain confidence level.}
}
}
