% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax.fun}
\alias{softmax.fun}
\title{Adjusted softmax transformation function}
\usage{
softmax.fun(msis)
}
\arguments{
\item{x}{a dataframe with \code{n} rows and \code{F} columns, where \code{n} 
is the number of observations and \code{F} is the number of candidate methods.}
}
\description{
In mathematics, softmax function maps a K-dimensional 
arbitrary real vector to another K-dimensional probability vector. 
Based on the properties of softmax function, the fitted values 
obtained from the pre-trained algorithm for each time series 
are transformed into probabilities. According to the adjusted 
softmax function, the method with a smaller score value 
in the candidate methods for each time series has a 
larger probability than the other methods. 
Consequently, the appropriate methods for each time series 
can be picked based on the obtained probabilities.
}
