% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_generate.R
\name{ts_generate}
\alias{ts_generate}
\title{Generate time series with various length by GRATIS.}
\usage{
ts_generate(
  n.ts = 1,
  freq = 1,
  nComp = NULL,
  length = NULL,
  h = 1,
  parallel = FALSE,
  num.cores = 2
)
}
\arguments{
\item{n.ts}{number of time series to be generated.}

\item{freq}{seasonal period of time series to be generated. For example, 
\code{freq} of the yearly, quarterly, and monthly series are 1, 4, and 12, respectively.}

\item{nComp}{number of mixing components when simulating time series using MAR models.}

\item{length}{length vector of time series to be generated. The length of 
the generated time series is simply sampled from the length vector.}

\item{h}{forecasting horizon of the generated time series.}

\item{parallel}{logical. If TRUE then the generation process is conducted in parallel.}

\item{num.cores}{the specified amount of parallel processes to be used if parallel = TRUE.}
}
\value{
A list of the generated time series.
}
\description{
Generate time series with certain seasonal period and length according to 
GRATIS proposed by Kang et al. (2019).
}
\examples{
 x_y <- ts_generate(n.ts = 10, freq = 4, nComp = 2, length = c(20,25,30), h = 8)
 x_y$Q1$pars
 forecast::autoplot(x_y$Q1$x)
}
\references{
Kang et al. (2019).
}
